+++ lib/resolvefilter.go
@@ -1,6 +1,12 @@
 package sous
 
-import "fmt"
+import (
+	"fmt"
+	"log"
+	"runtime/debug"
+
+	"github.com/samsalisbury/semv"
+)
 
 type (
 	// A ResolveFilter filters Deployments, DeployStates and Clusters for the
@@ -8,22 +14,41 @@ type (
 	ResolveFilter struct {
 		Repo     string
 		Offset   ResolveFieldMatcher
-		Tag      string
+		Tag      ResolveFieldMatcher
 		Revision string
 		Flavor   ResolveFieldMatcher
 		Cluster  string
 		Status   DeployStatus
 	}
 
-	// A ResolveFieldMatcher matches against any particular string, or all strings.
+	// A ResolveFieldMatcher matches against any particular string, or All strings.
 	ResolveFieldMatcher struct {
-		All   bool
-		Match string
+		Match *string
 	}
 )
 
+// NewResolveFieldMatcher wraps a string in a ResolveFieldMatcher that matches that string.
+func NewResolveFieldMatcher(match string) ResolveFieldMatcher {
+	return ResolveFieldMatcher{Match: &match}
+}
+
+// All returns true if this Matcher matches all values, or false if it matches
+// a specific value.
+func (matcher ResolveFieldMatcher) All() bool {
+	return matcher.Match == nil
+}
+
 func (matcher ResolveFieldMatcher) match(against string) bool {
-	return matcher.All || against == matcher.Match
+	return matcher.All() || against == *matcher.Match
+}
+
+// ValueOr returns the match value for this matcher, or the default value
+// provided if the matcher matches all values.
+func (matcher ResolveFieldMatcher) ValueOr(def string) string {
+	if matcher.All() {
+		return def
+	}
+	return *matcher.Match
 }
 
 func (rf *ResolveFilter) matchRepo(repo string) bool {
@@ -35,7 +60,7 @@ func (rf *ResolveFilter) matchOffset(offset string) bool {
 }
 
 func (rf *ResolveFilter) matchTag(tag string) bool {
-	return rf.Tag == "" || tag == rf.Tag
+	return rf.Tag.match(tag)
 }
 
 func (rf *ResolveFilter) matchRevision(rev string) bool {
@@ -54,13 +79,26 @@ func (rf *ResolveFilter) matchDeployStatus(status DeployStatus) bool {
 	return (rf.Status == DeployStatusAny || status == rf.Status)
 }
 
-// All returns true if the ResolveFilter would allow all deployments.
+// SetTag sets the tag based on a tag string - is ensures the tag parses as semver.
+func (rf *ResolveFilter) SetTag(tag string) error {
+	tagVersion, err := parseSemverTagWithOptionalPrefix(tag)
+	if err != nil {
+		log.Printf("%q", tag)
+		log.Print(string(debug.Stack()))
+		return fmt.Errorf("version %q not valid: expected something like [servicename-]1.2.3", tag)
+	}
+
+	rf.Tag = NewResolveFieldMatcher(tagVersion.Format(semv.Complete))
+	return nil
+}
+
+// All returns true if the ResolveFilter would allow All deployments.
 func (rf *ResolveFilter) All() bool {
 	return rf.Repo == "" &&
-		rf.Offset.All &&
-		rf.Tag == "" &&
+		rf.Offset.All() &&
+		rf.Tag.All() &&
 		rf.Revision == "" &&
-		rf.Flavor.All &&
+		rf.Flavor.All() &&
 		rf.Cluster == ""
 }
 
@@ -68,24 +106,24 @@ func (rf *ResolveFilter) All() bool {
 // describes a complete specific source location (i.e. it has exact Repo and
 // Offset matches set). Otherwise it returns a zero SourceLocation and false.
 func (rf *ResolveFilter) SourceLocation() (SourceLocation, bool) {
-	if rf.Repo == "*" || rf.Repo == "" || rf.Offset.All {
+	if rf.Repo == "*" || rf.Repo == "" || rf.Offset.All() {
 		return SourceLocation{}, false
 	}
 	return SourceLocation{
 		Repo: rf.Repo,
-		Dir:  rf.Offset.Match,
+		Dir:  *rf.Offset.Match,
 	}, true
 }
 
 // SourceID returns a SourceID based on the ResolveFilter and a ManifestID
 func (rf *ResolveFilter) SourceID(mid ManifestID) (SourceID, error) {
-	if rf.Tag == "" {
+	if rf.Tag.All() {
 		return SourceID{}, fmt.Errorf("you must provide the -tag flag")
 	}
 
-	newVersion, err := parseSemverTagWithOptionalPrefix(rf.Tag)
+	newVersion, err := semv.Parse(*rf.Tag.Match)
 	if err != nil {
-		return SourceID{}, fmt.Errorf("version %q not valid: expected something like [servicename-]1.2.3", rf.Tag)
+		return SourceID{}, err
 	}
 
 	return mid.Source.SourceID(newVersion), nil
@@ -100,26 +138,23 @@ func (rf *ResolveFilter) DeploymentID(mid ManifestID) (DeploymentID, error) {
 }
 
 func (rf *ResolveFilter) String() string {
-	cl, fl, rp, of, tg, rv := rf.Cluster, rf.Flavor.Match, rf.Repo, rf.Offset.Match, rf.Tag, rf.Revision
+	cl, rp, rv := rf.Cluster, rf.Repo, rf.Revision
 	if cl == "" {
 		cl = `*`
-	}
-	if rf.Flavor.All {
-		fl = `*`
 	}
 	if rp == "" {
 		rp = `*`
 	}
-	if rf.Offset.All {
-		of = `*`
-	}
-	if tg == "" {
-		tg = `*`
-	}
 	if rv == "" {
 		rv = `*`
 	}
-	return fmt.Sprintf("<cluster:%s repo:%s offset:%s flavor:%s tag:%s revision:%s>",
+
+	fl := rf.Flavor.ValueOr("*")
+	of := rf.Offset.ValueOr("*")
+	tg := rf.Tag.ValueOr("*")
+
+	return fmt.Sprintf(
+		"<cluster:%s repo:%s offset:%s flavor:%s tag:%s revision:%s>",
 		cl, rp, of, fl, tg, rv)
 }